ai_study_coach/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
